feat: add neuro net module
This commit introduces the Neuron and DenseLayer classes to the new module. The Neuron class represents a single neuron in a neural network and provides the ability to compute the output of the neuron given the input and weights. The DenseLayer class represents a dense layer in a neural network and includes a collection of neurons. It allows for the computation of the layer's output given the inputs or the output of the previous layer. These additions enhance the functionality of the module for working with neural networks. This commit adds also __init_
file for nnkit package.